ParticleRenderer class
